requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.9.1
chardet==3.0.4
coreapi==2.1.1
coreschema==0.0.4
cryptography==2.0.2
decorator==4.0.11
Django==1.10.5
django-colorfield==0.1.11
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-grappelli==2.9.1
django-rest-swagger==2.1.1
djangorestframework==3.5.3
Fabric3==1.13.1.post1
flake8==3.2.1
idna==2.5
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.5.3
olefile==0.44
openapi-codec==1.2.1
paramiko==2.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyasn1==0.1.9
pycodestyle==2.2.0
pycparser==2.17
pyflakes==1.3.0
Pygments==2.2.0
PyNaCl==1.1.2
pytz==2017.2
requests==2.13.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
